Conflicts:
	experiments/cufft_ntt_cmp/Makefile
	src/benchmark_polynomial.cpp

+++ src/distribution.h
@@ -1,3 +1,20 @@
+/**
+ * cuYASHE
+ * Copyright (C) 2015-2016 cuYASHE Authors
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 #ifndef DISTRIBUTION_H
 #define DISTRIBUTION_H
 
@@ -17,7 +34,7 @@ enum kind_t
 };
 
 #define MAX_DEGREE 16384
-#define SEED (unsigned long)(7226776555987513888)
+#define SEED (unsigned long long)(7226776555987513888)
 
 class Distribution{
   private:
